requirements.txt
discord.py>=2.0.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
jinja2>=3.1.0
pillow>=10.0.0
playwright>=1.40.0
